Kalman_pointtracker improvements that make the color presentation more consistent.  Needs more work with aging out but much better in maintaining the same color across generations.

+++ VB_Classes/Highlight.vb
@@ -6,7 +6,7 @@ Public Class Highlight_Basics
     Dim highlightRect As New cv.Rect
     Dim preKalmanRect As New cv.Rect
     Dim highlightMask As New cv.Mat
-    Public viewObjects As New SortedList(Of Integer, viewObject)(New compareAllowIdenticalIntInverted)
+    Public viewObjects As New SortedList(Of Single, viewObject)(New compareAllowIdenticalIntInverted)
     Public Sub New(ocvb As VBocvb)
         setCaller(ocvb)
         If standalone Then reduction = New Reduction_KNN_Color(ocvb)
